+++ CHANGELOG.md
@@ -14,6 +14,12 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 * ```Fixed``` for any bug fixes.
 * ```Security``` in case of vulnerabilities.
 
+## [2023.1.3.6] - 2023-03-06
+
+### Added
+
+- Unique indexes for xstock tables
+
 
 ## [2023.1.2.17] - 2023-02-17